fastapi-backend/app/xlsx_analysis/requirements.txt
cffi==1.17.1
colorclass==2.2.2
cryptography==44.0.0
easygui==0.98.3
et_xmlfile==2.0.0
msoffcrypto-tool==5.4.2
numpy==2.2.2
olefile==0.47
oletools==0.60.2
openpyxl==3.1.5
pandas==2.2.3
pcodedmp==1.2.6
pycparser==2.22
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-magic==0.4.27
pytz==2024.2
six==1.17.0
tzdata==2025.1
